+++ src/KernexClient.ts
@@ -1,4 +1,4 @@
-import KernexResource from './KernexResource';
+import KernexResource, {KernexResourceOptions} from './KernexResource';
 import { ResourcesMap } from './resources';
 
 export default class KernexClient<Resources extends ResourcesMap> {
@@ -7,13 +7,15 @@ export default class KernexClient<Resources extends ResourcesMap> {
   /**
    * Get an instance of resource class.
    * @param resourceName
+   * @param options
    */
   resource<
     ResourceName extends keyof Resources & string = string,
     Resource = Resources[ResourceName]
   >(
     resourceName: ResourceName,
+    options?: KernexResourceOptions,
   ) {
-    return new KernexResource<Resource>(this.baseUrl, this.apiKey, resourceName);
+    return new KernexResource<Resource>(this.baseUrl, this.apiKey, resourceName, options);
   }
 }